АДМ. Направления и пожарные зоны. При задании номера направлению/пожарной зоне более 65535, итерирования на единицу при создании нового направления/зоны не происходит. В охранных зонах такой проблемы нет. У МПТ, задержек и направлений та же беда.

+++ Projects/FireAdministrator/Modules/GroupControllerModule/Zones/ViewModels/ZoneDetailsViewModel.cs
@@ -29,7 +29,7 @@ namespace GKModule.ViewModels
 					Fire2Count = LastFire2Count
 				};
 				if (GKManager.Zones.Count != 0)
-					Zone.No = (ushort)(GKManager.Zones.Select(x => x.No).Max() + 1);
+					Zone.No = (GKManager.Zones.Select(x => x.No).Max() + 1);
 			}
 			else
 			{
